+++ dd/cudd.pyx
@@ -32,7 +32,6 @@ from libc.stdio cimport FILE, fdopen, fopen, fclose
 cimport libc.stdint as stdint
 from cpython cimport bool as python_bool
 from cpython.mem cimport PyMem_Malloc, PyMem_Free
-import psutil
 
 
 IF USE_CYSIGNALS:
@@ -388,11 +387,13 @@ cdef class BDD:
             # could in principle have an
             # arbitrary value when `__dealloc__`
             # is executed.
-        total_memory = psutil.virtual_memory().total
+        total_memory = _utils.total_memory()
         default_memory = DEFAULT_MEMORY
         if memory_estimate is None:
             memory_estimate = default_memory
-        if memory_estimate >= total_memory:
+        if total_memory is None:
+            pass
+        elif memory_estimate >= total_memory:
             msg = (
                 'Error in `dd.cudd`: '
                 'total physical memory '